研修用に微修正
