+++ .gitignore
@@ -27,3 +27,6 @@ replay_pid*
 .gradle/
 build/
 out/
+
+# Editor
+.idea/
