+++ src/main/java/jp/co/athub/todo/TodolistBootcampApplication.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package jp.co.athub.todo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
