+++ src/main/java/jp/co/athub/todo/WebController.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package jp.co.athub.todo;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
