+++ src/main/resources/templates/index.html
@@ -17,8 +17,8 @@
         <tbody>
             <tr th:each="todo : ${todos}">
                 <td th:text="${todo.title}"></td>
-                <td th:text="${todo.date} ?: ''"></td>
-                <td th:text="${todo.assignee} ?: ''"></td>
+                <td th:text=""></td>
+                <td th:text=""></td>
             </tr>
         </tbody>
     </table>
